// src/interfaces/participants.tsx
export interface RootParticipant {
  puuid?: string;
  assists: number;
  championId: number;
  championName: string;
  champLevel: number;
  deaths: number;
  goldEarned: number;
  goldSpent: number;
  item0: number;
  item1: number;
  item2: number;
  item3: number;
  item4: number;
  item5: number;
  item6: number;
  kills: number;
  summoner1Id: number;
  summoner2Id: number;
  summonerName: string;
  teamPosition: string;
  totalMinionsKilled: number;
  win: boolean;
}